Move getCurrencyInfo to the currency package
